+++ src/main/kotlin/blackjack/config/GameConfig.kt
@@ -1,7 +1,5 @@
 package blackjack.config
 
-import blackjack.domain.BlackJackOutcomeCalculator
-import blackjack.domain.GameOutcomeCalculator
 import blackjack.ui.GameController
 import blackjack.ui.GameInput
 import blackjack.ui.GameInputImpl
@@ -14,12 +12,8 @@ object GameConfig {
     val gameController: GameController by lazy {
         GameController(
             gameInput = gameInput,
-            gameOutput = gameOutput,
-            gameOutcomeCalculator = gameOutcomeCalculator
+            gameOutput = gameOutput
         )
-    }
-    private val gameOutcomeCalculator: GameOutcomeCalculator by lazy {
-        BlackJackOutcomeCalculator()
     }
     private val gameOutput: GameOutput by lazy { GameOutputImpl }
     private val gameInput: GameInput by lazy { GameInputProxy(target = GameInputImpl) }
