+++ src/main/kotlin/blackjack/domain/GameOutcomeCalculator.kt
@@ -1,5 +1,5 @@
 package blackjack.domain
 
 fun interface GameOutcomeCalculator {
-    fun calculate(dealer: Dealer, players: Players): GameResult
+    fun calculate(players: Players): GameResult
 }
