+++ src/main/kotlin/blackjack/domain/Gamer.kt
@@ -1,21 +1,14 @@
 package blackjack.domain
 
-class Gamer(override val name: String, override val deck: Deck = Deck()) : Player {
-    private val interDeck = deck.copy()
-
-    override fun calculateScore(): Int = interDeck.score()
-
+class Gamer(name: String, deck: Deck = Deck()) : Player(
+    name = name,
+    deck = deck.copy()
+) {
     override fun addCard(card: Card) {
         check(isAddable()) {
             "현재 점수가 21점 이상이기에 카드를 추가할 수 없습니다."
-        }
-
-        interDeck.add(card)
+        }.run { super.addCard(card) }
     }
 
-    override fun addCardAll(values: Collection<Card>) = values.forEach(::addCard)
-
-    override fun isAddable(): Boolean = interDeck.score() < Game.THRESHOLD
-
-    override fun currentDeck(): Deck = interDeck.copy()
+    override fun isAddable(): Boolean = deck.score() < Game.THRESHOLD
 }
