+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,12 +1,20 @@
 package blackjack.domain
 
-sealed interface Player {
-    val name: String
-    val deck: Deck
+sealed class Player(
+    val name: String,
+    deck: Deck
+) {
+    val deck: Deck = deck.copy()
 
-    fun addCard(card: Card)
-    fun addCardAll(values: Collection<Card>)
-    fun isAddable(): Boolean
-    fun calculateScore(): Int
-    fun currentDeck(): Deck
+    open fun addCard(card: Card) {
+        deck.add(card)
+    }
+
+    fun addCardAll(values: Collection<Card>) = values.forEach(::addCard)
+
+    fun currentDeck(): Deck = deck.copy()
+
+    fun calculateScore(): Int = deck.score()
+
+    abstract fun isAddable(): Boolean
 }
